+++ src/Core/Algorithms/Legacy/Fields/CMakeLists.txt
@@ -102,6 +102,8 @@ SET(Core_Algorithms_Legacy_Fields_HEADERS
   MeshData/GetMeshQualityFieldAlgo.h
   Cleanup/RemoveUnusedNodes.h
   Cleanup/CleanupTetMesh.h
+  MeshDerivatives/CalculateMeshCenterAlgo.h
+
 )
 
 SET(Core_Algorithms_Legacy_Fields_SRCS
@@ -185,7 +187,7 @@ SET(Core_Algorithms_Legacy_Fields_SRCS
   #MeshData/GetSurfaceElemNormals.cc
   MeshData/GetMeshQualityFieldAlgo.cc
   #MeshDerivatives/CalculateMeshConnector.cc
-  #MeshDerivatives/CalculateMeshCenter.cc
+  MeshDerivatives/CalculateMeshCenterAlgo.cc
   #MeshDerivatives/GetCentroids.cc
   MeshDerivatives/GetFieldBoundaryAlgo.cc
   #MeshDerivatives/GetBoundingBox.cc
@@ -213,6 +215,7 @@ SET(Core_Algorithms_Legacy_Fields_SRCS
   #TransformMesh/PolarProjectionTools.cc
   #TransformMesh/GeneratePolarProjectionImage.cc
   #TracePoints/TracePoints.cc
+  
 )
 
 SCIRUN_ADD_LIBRARY(Core_Algorithms_Legacy_Fields
