+++ src/Modules/Fields/CMakeLists.txt
@@ -33,7 +33,6 @@ SET(Modules_Fields_SRCS
   EditMeshBoundingBox.cc
   CalculateNodeLocationFrequency.cc
   RefineTetMeshLocally.cc
-  CalculateMeshCenter.cc
 )
 
 SET(Modules_Fields_HEADERS
@@ -44,7 +43,6 @@ SET(Modules_Fields_HEADERS
   CalculateNodeLocationFrequency.h
   RefineTetMeshLocally.h
   share.h
-  CalculateMeshCenter.h
 )
 
 SCIRUN_ADD_LIBRARY(Modules_Fields
