+++ src/Modules/Legacy/Fields/CalculateMeshCenter.cc
@@ -1,81 +1,65 @@
 /*
- * The MIT License
- * 
- * Copyright (c) 2015 Scientific Computing and Imaging Institute,
- * University of Utah.
- * 
- * 
- * Permission is hereby granted, free of charge, to any person obtaining a
- * copy of this software and associated documentation files (the "Software"),
- * to deal in the Software without restriction, including without limitation
- * the rights to use, copy, modify, merge, publish, distribute, sublicense,
- * and/or sell copies of the Software, and to permit persons to whom the
- * Software is furnished to do so, subject to the following conditions:
- * 
- * The above copyright notice and this permission notice shall be included
- * in all copies or substantial portions of the Software.
- * 
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS
- * OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL
- * THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
- * FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
- * DEALINGS IN THE SOFTWARE.
- */
-/// @todo Documentation Modules/Legacy/Fields/CalculateMeshCenter.cc
+   For more information, please see: http://software.sci.utah.edu
 
-#include <Core/Algorithms/Fields/MeshDerivatives/CalculateMeshCenter.h>
+   The MIT License
 
-#include <Core/Datatypes/Field.h>
+   Copyright (c) 2015 Scientific Computing and Imaging Institute,
+   University of Utah.
 
-#include <Dataflow/Network/Ports/FieldPort.h>
-#include <Dataflow/Network/Module.h>
+   Permission is hereby granted, free of charge, to any person obtaining a
+   copy of this software and associated documentation files (the "Software"),
+   to deal in the Software without restriction, including without limitation
+   the rights to use, copy, modify, merge, publish, distribute, sublicense,
+   and/or sell copies of the Software, and to permit persons to whom the
+   Software is furnished to do so, subject to the following conditions:
+
+   The above copyright notice and this permission notice shall be included
+   in all copies or substantial portions of the Software.
+
+   THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS
+   OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+   FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL
+   THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+   LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
+   FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
+   DEALINGS IN THE SOFTWARE.
+*/
 
 
-namespace SCIRun {
+#include <Modules/Legacy/Fields/CalculateMeshCenter.h>
+#include <Core/Algorithms/Base/AlgorithmBase.h>
+#include <Core/Datatypes/Legacy/Field/Field.h>
+#include <Core/Datatypes/Scalar.h>
+#include <Core/Algorithms/Legacy/Fields/MeshDerivatives/CalculateMeshCenterAlgo.h>
 
-using namespace SCIRun;
+using namespace SCIRun::Modules::Fields;
+using namespace SCIRun::Core::Datatypes;
+using namespace SCIRun::Dataflow::Networks;
+using namespace SCIRun::Core::Algorithms::Fields;
 
-class CalculateMeshCenter : public Module {
-  public:
-    CalculateMeshCenter(GuiContext*);
-    virtual ~CalculateMeshCenter() {}
+MODULE_INFO_DEF(CalculateMeshCenter, NewField, SCIRun)
 
-    virtual void execute();
-
-  private:
-    SCIRunAlgo::CalculateMeshCenterAlgo algo_;
-    GuiString method_;
-};
-
-
-DECLARE_MAKER(CalculateMeshCenter)
-
-CalculateMeshCenter::CalculateMeshCenter(GuiContext* ctx) :
-  Module("CalculateMeshCenter", ctx, Source, "MiscField", "SCIRun"),
-  method_(get_ctx()->subVar("method"),"weighted-elem-center")
+CalculateMeshCenter::CalculateMeshCenter() : Module(staticInfo_)
 {
-  algo_.set_progress_reporter(this);
+  INITIALIZE_PORT(InputField);
+  INITIALIZE_PORT(OutputField);
 }
 
-
-void
-CalculateMeshCenter::execute()
+void CalculateMeshCenter::setStateDefaults()
 {
-  FieldHandle input, output;
-  
-  get_input_handle("Input",input,true);
-  
-  if (inputs_changed_ || method_.changed() || !oport_cached("Output"))
+  setStateStringFromAlgoOption(Parameters::Method);
+}
+
+void CalculateMeshCenter::execute()
+{
+  auto field = getRequiredInput(InputField);
+
+  if (needToExecute())
   {
-    algo_.set_option("method",method_.get());
-    if (!(algo_.run(input,output))) return;
-    send_output_handle("Output",output,true);
+    setAlgoOptionFromState(Parameters::Method);
+    
+    auto output = algo().run(withInputData((InputField, field)));
+
+    sendOutputFromAlgorithm(OutputField,output);
   }
-
 }
-
-} // End namespace SCIRun
-
-
